Added frontend for get and put requests.
