+++ Backend/DTOs/AlbumDTOs/AlbumSummaryDTO.cs
@@ -1,0 +1,10 @@
+﻿namespace MusicCollectionAPI.Backend.DTOs.AlbumDTOs
+{
+    public record class AlbumSummaryDTO(
+        int Id,
+        string Title,
+        string Artist,
+        string Genre,
+        DateOnly ReleaseDate,
+        string Format);
+}
