+++ Backend/DTOs/FormatDTOs/FormatDTO.cs
@@ -1,0 +1,6 @@
+﻿namespace MusicCollectionAPI.Backend.DTOs.FormatDTOs
+{
+    public record class FormatDTO(
+        int Id,
+        string Name);
+}
