+++ Backend/DTOs/FormatDTOs/UpdateFormatDTO.cs
@@ -1,0 +1,9 @@
+﻿using System.ComponentModel.DataAnnotations;
+
+namespace MusicCollectionAPI.Backend.DTOs.FormatDTOs
+{
+    public record class UpdateFormatDTO
+    (
+        [Required][StringLength(30)] string Name
+    );
+}
