+++ Backend/DTOs/GenreDTOs/GenreDTO.cs
@@ -1,0 +1,6 @@
+﻿namespace MusicCollectionAPI.Backend.DTOs.GenreDTOs
+{
+    public record class GenreDTO(
+        int Id,
+        string Name);
+}
