+++ Backend/Entities/Format.cs
@@ -1,0 +1,8 @@
+﻿namespace MusicCollectionAPI.Backend.Entities
+{
+    public class Format
+    {
+        public int Id { get; set; }
+        public required string Name { get; set; }
+    }
+}
