+++ Backend/Entities/Genre.cs
@@ -1,0 +1,8 @@
+﻿namespace MusicCollectionAPI.Backend.Entities
+{
+    public class Genre
+    {
+        public int Id { get; set; }
+        public required string Name { get; set; }
+    }
+}
