--- DTOs/AlbumDTOs/AlbumDetailsDTO.cs
@@ -1,10 +1,0 @@
-﻿namespace MusicCollectionAPI.DTOs.AlbumDTOs
-{
-    public record class AlbumDetailsDTO(
-        int Id,
-        string Title,
-        string Artist,
-        int GenreId,
-        DateOnly ReleaseDate,
-        int FormatId);
-}
