--- DTOs/AlbumDTOs/CreateAlbumDTO.cs
@@ -1,13 +1,0 @@
-﻿using System.ComponentModel.DataAnnotations;
-
-namespace MusicCollectionAPI.DTOs.AlbumDTOs
-{
-    public record class CreateAlbumDTO
-    (
-        [Required][StringLength(80)] string Title,
-        [Required][StringLength(80)] string Artist,
-        int GenreId,
-        DateOnly ReleaseDate,
-        int FormatId
-    );
-}
