--- DTOs/FormatDTOs/FormatDTO.cs
@@ -1,6 +1,0 @@
-﻿namespace MusicCollectionAPI.DTOs.FormatDTOs
-{
-    public record class FormatDTO(
-        int Id,
-        string Name);
-}
