--- DTOs/GenreDTOs/CreateGenreDTO.cs
@@ -1,9 +1,0 @@
-﻿using System.ComponentModel.DataAnnotations;
-
-namespace MusicCollectionAPI.DTOs.GenreDTOs
-{
-    public record class CreateGenreDTO
-    (
-        [Required][StringLength(50)] string Name
-    );
-}
