--- DTOs/GenreDTOs/GenreDTO.cs
@@ -1,6 +1,0 @@
-﻿namespace MusicCollectionAPI.DTOs.GenreDTOs
-{
-    public record class GenreDTO(
-        int Id,
-        string Name);
-}
