--- DTOs/GenreDTOs/UpdateGenreDTO.cs
@@ -1,9 +1,0 @@
-﻿using System.ComponentModel.DataAnnotations;
-
-namespace MusicCollectionAPI.DTOs.GenreDTOs
-{
-    public record class UpdateGenreDTO
-    (
-        [Required][StringLength(50)] string Name
-    );
-}
