--- Entities/Format.cs
@@ -1,8 +1,0 @@
-﻿namespace MusicCollectionAPI.Entities
-{
-    public class Format
-    {
-        public int Id { get; set; }
-        public required string Name { get; set; }
-    }
-}
