--- Entities/Genre.cs
@@ -1,8 +1,0 @@
-﻿namespace MusicCollectionAPI.Entities
-{
-    public class Genre
-    {
-        public int Id { get; set; }
-        public required string Name { get; set; }
-    }
-}
