+++ Frontend/index.html
@@ -1,0 +1,38 @@
+﻿<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Music Colleciton API</title>
+    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css">
+</head>
+<body class="container mt-5">
+    <h1 class="mb-4">Music Collection API</h1>
+    <form id="albumForm">
+        <div class="mb-2">
+            <label for="title" class="form-label">Album Title</label>
+            <input type="text" class="form-control" id="title" required>
+        </div>
+        <div class="mb-2">
+            <label for="artist" class="form-label">Aritst</label>
+            <input type="text" class="form-control" id="artist" required>
+        </div>
+        <div class="mb-2">
+            <label for="genreId" class="form-label">Genre ID</label>
+            <input type="number" class="form-control" id="genreId" required>
+        </div>
+        <div class="mb-2">
+            <label for="releaseDate" class="form-label">Release Date</label>
+            <input type="date" class="form-control" id="releaseDate" required>
+        </div>
+        <div class="mb-3">
+            <label for="formatId" class="form-label">Format ID</label>
+            <input type="number" class="form-control" id="formatId" required>
+        </div>
+        <button id="submit-button" type="submit" class="btn btn-primary">Add Album</button>
+    </form>
+    <h2 class="mt-4">Album List</h2>
+    <ul id="albumsList" class="list-group mt-3"></ul>
+    <script src="app.js"></script>
+</body>
+</html>
